pyproject.toml
[project]
name = "SciDef"
version = "0.1.0"
description = "**SciDef** provides resources to support research on **definition extraction and definition similarity** from scientific literature."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.12.0",
    "argparse>=1.4.0",
    "asyncio>=3.4.3",
    "config>=0.5.1",
    "datasets>=3.6.0",
    "dspy>=3.0.3",
    "grobid-client-python>=0.1.2",
    "huggingface-hub>=0.33.2",
    "kagglehub>=0.3.13",
    "logging>=0.4.9.6",
    "lxml>=6.0.2",
    "openai>=1.93.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.53.0",
]

[dependency-groups]
dev = [
    "flashinfer-python>=0.5.2",
    "hf-transfer>=0.1.9",
    "ipykernel>=6.30.1",
    "streamlit>=1.51.0",
    "vllm>=0.11.0",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
    "weave>=0.52.20",
    "torch-c-dlpack-ext>=0.1.4",
    "timm>=1.0.22",
    "bitsandbytes>=0.49.0",
    "ty>=0.0.2",
    "nvidia-modelopt>=0.40.0",
    "ruff>=0.12.1",
    "pre-commit>=4.3.0",
    "marimo>=0.14.10",
    "label-studio>=1.22.0",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
ignore = ["C901"]
